refactor(data): move SFT assistant reconstruction out of CT
Keep Continuous Token builders focused on runtime token construction and merging. Move gold assistant-message reconstruction into a package-local SFT adapter while preserving every model-family behavior carried by this branch.

Split reconstruction tests from the runtime suite and retain MRO coverage for text and VL builders. The migration matches all 122 recorded token/result/error behaviors from the pre-migration baseline.
