+++ verl/utils/tokenizer/continuous_token.py
@@ -99,17 +99,13 @@ class ContinuousTokenBuilder:
         appends model-generated assistant tokens, and ``align_response_metadata``
         applies the recorded token edits to masks/logprobs.
 
-    SFT-facing API:
-        ``tokenize_assistant_message`` converts one prepared gold assistant
-        message into the same continuation-token shape produced by rollout.
-
     Developer extension APIs:
         Model-specific builders should subclass this class and keep the runtime
         API contracts above stable. Chat template specific behavior belongs in hooks
         such as ``_tokenize_tool_group``, ``_tokenize_single_non_tool``,
         ``_should_fuse_generation_prompt_with_last_group``,
-        ``_tokenize_generation_prompt_delta``, ``_assistant_terminator_ids``, and
-        ``_merge_non_assistant_token_ids``. ``render_delta_token_id`` is the shared
+        ``_tokenize_generation_prompt_delta`` and ``_merge_non_assistant_token_ids``.
+        ``render_delta_token_id`` is the shared
         suffix-diff helper those hooks can reuse.
     """
 
@@ -216,125 +212,6 @@ class ContinuousTokenBuilder:
             kind="assistant",
         )
 
-    def tokenize_assistant_message(
-        self,
-        message: dict[str, Any],
-        *,
-        tools: list[dict[str, Any]] | None = None,
-        previous_messages: list[dict[str, Any]] | None = None,
-    ) -> list[int]:
-        """Encode one prepared SFT assistant message into rollout-shaped tokens.
-
-        Rollout engines already provide generated assistant token IDs and should
-        continue to call :meth:`merge_assistant_tokens` directly. SFT datasets,
-        however, start from structured assistant messages. This helper renders
-        the assistant message exactly once behind a fixed synthetic prompt and
-        removes that prompt at the token boundary. No token from the real
-        trajectory prefix is decoded or re-encoded.
-
-        Chat templates may emit whitespace after the assistant stop token even
-        though a rollout server stops at that token. The returned suffix is
-        therefore normalized to the runtime stop shape before it is merged.
-        """
-        del previous_messages
-        if message.get("role") != "assistant":
-            raise ValueError(
-                f"Continuous Token assistant encoding requires role='assistant', got {message.get('role')!r}"
-            )
-
-        rendered_message = self._prepare_assistant_message_for_render(message)
-        synthetic_prompt = [_SYNTHETIC_SYSTEM_MESSAGE, _SYNTHETIC_USER_MESSAGE]
-        prompt_text = self._render_text(synthetic_prompt, add_generation_prompt=True, tools=tools)
-        completed_text = self._render_text(
-            [*synthetic_prompt, rendered_message],
-            add_generation_prompt=False,
-            tools=tools,
-        )
-        if not completed_text.startswith(prompt_text):
-            raise ValueError(
-                "Continuous Token assistant encoding requires the generation prompt to be a text prefix "
-                "of the completed assistant turn"
-            )
-
-        # Encode only the model-produced continuation. Encoding prompt+response
-        # together would let the tokenizer merge across their boundary and would
-        # no longer match rollout, where the prompt tokens already exist.
-        assistant_text = completed_text[len(prompt_text) :]
-        assistant_token_ids = normalize_token_ids(self.tokenizer.encode(assistant_text, add_special_tokens=False))
-        if not assistant_token_ids:
-            raise ValueError("Continuous Token assistant encoding produced an empty token-id suffix")
-        return self._normalize_assistant_token_ids(assistant_token_ids, message)
-
-    def _prepare_assistant_message_for_render(self, message: dict[str, Any]) -> dict[str, Any]:
-        """Normalize one prepared message before extracting its generated text."""
-        return message
-
-    def _render_assistant_turn_deltas(
-        self,
-        message: dict[str, Any],
-        *,
-        tools: list[dict[str, Any]] | None = None,
-    ) -> tuple[str, str]:
-        """Render generation/completion deltas behind one fixed context.
-
-        Unlike :meth:`tokenize_assistant_message`, this helper does not require
-        the generation prompt to prefix the completed turn. Model builders with
-        template-specific generation scaffolds can compare the two deltas and
-        reconstruct the rollout continuation explicitly.
-        """
-        synthetic_prompt = [_SYNTHETIC_SYSTEM_MESSAGE, _SYNTHETIC_USER_MESSAGE]
-        context_text = self._render_text(synthetic_prompt, add_generation_prompt=False, tools=tools)
-        prompt_text = self._render_text(synthetic_prompt, add_generation_prompt=True, tools=tools)
-        completed_text = self._render_text(
-            [*synthetic_prompt, message],
-            add_generation_prompt=False,
-            tools=tools,
-        )
-        if not prompt_text.startswith(context_text) or not completed_text.startswith(context_text):
-            raise ValueError(
-                "Continuous Token assistant encoding requires both generation and completed renders "
-                "to preserve the fixed synthetic context"
-            )
-        return prompt_text[len(context_text) :], completed_text[len(context_text) :]
-
-    def _normalize_assistant_token_ids(
-        self,
-        assistant_token_ids: list[int],
-        message: dict[str, Any],
-    ) -> list[int]:
-        """Trim template-only tokens after the final generated stop token."""
-        terminator_ids = self._assistant_terminator_ids(message)
-        if not terminator_ids:
-            return list(assistant_token_ids)
-        # A rollout server stops as soon as it generates any configured stop
-        # token. Mirror that first-stop behavior when reconstructing gold SFT
-        # continuations, including the uncommon case where the serialized
-        # content itself contains a special terminator token.
-        for index, token_id in enumerate(assistant_token_ids):
-            if token_id in terminator_ids:
-                return list(assistant_token_ids[: index + 1])
-        raise ValueError(
-            "Continuous Token assistant token-id suffix does not contain an accepted terminator "
-            f"{sorted(terminator_ids)}; tail={assistant_token_ids[-16:]}"
-        )
-
-    def _assistant_terminator_ids(self, message: dict[str, Any]) -> set[int]:
-        """Return valid generation stop tokens for one assistant message.
-
-        Most model families terminate every assistant continuation with EOS.
-        Builders whose protocol has message-dependent stops (for example, a
-        distinct tool-call terminator) override this hook.
-        """
-        del message
-        eos_token_id = getattr(self.tokenizer, "eos_token_id", None)
-        if isinstance(eos_token_id, int):
-            return {eos_token_id}
-        elif isinstance(eos_token_id, list | tuple | set):
-            return {int(token_id) for token_id in eos_token_id if token_id is not None}
-        elif eos_token_id is None:
-            return set()
-        raise TypeError(f"Unsupported eos_token_id type: {type(eos_token_id)!r}")
-
     def _merge_non_assistant_token_ids(
         self, runtime_token_ids: list[int], appended_token_ids: list[int]
     ) -> MergeResult:
@@ -366,25 +243,6 @@ class ContinuousTokenBuilder:
             **self.chat_template_kwargs,
         )
         return normalize_token_ids(tokenized)
-
-    def _render_text(
-        self,
-        messages: list[dict[str, Any]],
-        *,
-        add_generation_prompt: bool,
-        tools: list[dict[str, Any]] | None = None,
-    ) -> str:
-        rendered = apply_chat_template(
-            self.tokenizer,
-            messages,
-            tokenize=False,
-            add_generation_prompt=add_generation_prompt,
-            tools=tools,
-            **self.chat_template_kwargs,
-        )
-        if not isinstance(rendered, str):
-            raise TypeError(f"Expected chat template to render str, got {type(rendered).__name__}")
-        return rendered
 
     def render_delta_token_id(
         self,
@@ -616,24 +474,6 @@ class GptOssContinuousTokenBuilder(ContinuousTokenBuilder):
         # suffix diff, so keep generation-prompt derivation as a separate step.
         return False
 
-    def _prepare_assistant_message_for_render(self, message: dict[str, Any]) -> dict[str, Any]:
-        # Arrow-backed SFT datasets can materialize absent optional struct
-        # fields as null. Harmony checks key presence for ``tool_calls`` and
-        # ``thinking``, so remove null optionals and normalize nullable content
-        # before invoking the template. Keep this normalization in the builder
-        # as well as the dataset boundary for direct SFT callers.
-        rendered_message = {key: value for key, value in message.items() if value is not None}
-        if message.get("content") is None:
-            rendered_message["content"] = ""
-        return rendered_message
-
-    def _assistant_terminator_ids(self, message: dict[str, Any]) -> set[int]:
-        # Harmony final answers stop at <|return|> (the tokenizer EOS), while
-        # assistant tool calls stop at <|call|> and contain no EOS token.
-        if message.get("tool_calls"):
-            return {require_token_id(self.tokenizer, "<|call|>")}
-        return super()._assistant_terminator_ids(message)
-
     def _tokenize_tool_group(
         self,
         tool_messages: list[dict[str, Any]],
@@ -678,42 +518,6 @@ class QwenContinuousTokenBuilder(ContinuousTokenBuilder):
             raise ValueError(f"Expected Qwen newline to tokenize to one token, got {newline_ids!r}")
         self._newline_id = int(newline_ids[0])
         self._im_end_id = require_token_id(tokenizer, "<|im_end|>")
-
-    def _prepare_assistant_message_for_render(self, message: dict[str, Any]) -> dict[str, Any]:
-        """Preserve literal nested think tags in raw Qwen assistant output.
-
-        Qwen's template uses the last opening think tag when it infers
-        reasoning_content from content, which drops text when reasoning itself
-        mentions a literal think tag. Split only the outermost leading block and
-        pass the fields explicitly so every nested/generated tag survives.
-        """
-        enable_thinking = self.chat_template_kwargs.get("enable_thinking")
-        if isinstance(message.get("reasoning_content"), str):
-            if enable_thinking is not False:
-                return message
-            rendered_message = dict(message)
-            rendered_message["reasoning_content"] = ""
-            return rendered_message
-        content = message.get("content")
-        content_is_text_blocks = isinstance(content, list) and all(
-            isinstance(block, dict) and block.get("type") == "text" for block in content
-        )
-        if content_is_text_blocks:
-            content_text = "".join(str(block.get("text", "")) for block in content)
-        elif isinstance(content, str):
-            content_text = content
-        else:
-            return message
-        if not content_text.startswith("<think>") or "</think>" not in content_text:
-            return message
-
-        reasoning_content, answer_content = content_text[len("<think>") :].split("</think>", 1)
-        rendered_message = dict(message)
-        rendered_message["reasoning_content"] = reasoning_content if enable_thinking is not False else ""
-        rendered_message["content"] = (
-            [{"type": "text", "text": answer_content}] if content_is_text_blocks else answer_content
-        )
-        return rendered_message
 
     def _merge_non_assistant_token_ids(
         self, runtime_token_ids: list[int], appended_token_ids: list[int]
@@ -791,27 +595,6 @@ class MiniMaxText01ContinuousTokenBuilder(ContinuousTokenBuilder):
             rendered = rendered[: -len(self._generation_scaffold_text)]
         return rendered
 
-    def tokenize_assistant_message(
-        self,
-        message: dict[str, Any],
-        *,
-        tools: list[dict[str, Any]] | None = None,
-        previous_messages: list[dict[str, Any]] | None = None,
-    ) -> list[int]:
-        # Tool declarations sit after messages in the Text-01 template and do
-        # not affect the generated continuation. Omitting them from this fixed
-        # probe prevents their position from invalidating the prefix boundary.
-        del tools
-        return super().tokenize_assistant_message(
-            _prepare_minimax_legacy_assistant_message(message),
-            tools=None,
-            previous_messages=previous_messages,
-        )
-
-    def _assistant_terminator_ids(self, message: dict[str, Any]) -> set[int]:
-        del message
-        return {self._eos_id}
-
     def _should_fuse_generation_prompt_with_last_group(self) -> bool:
         # Text-01 append hooks encode role payloads directly, while the next
         # assistant header is a fixed scaffold appended by the hook below.
@@ -882,67 +665,6 @@ class MiniMaxContinuousTokenBuilder(ContinuousTokenBuilder):
         self._newline_id = int(newline_ids[0])
         self._eos_id = require_token_id(tokenizer, "[e~[")
 
-    def tokenize_assistant_message(
-        self,
-        message: dict[str, Any],
-        *,
-        tools: list[dict[str, Any]] | None = None,
-        previous_messages: list[dict[str, Any]] | None = None,
-    ) -> list[int]:
-        """Reconstruct the response behind MiniMax's always-open think prompt."""
-        del previous_messages
-        if message.get("role") != "assistant":
-            raise ValueError(
-                f"Continuous Token assistant encoding requires role='assistant', got {message.get('role')!r}"
-            )
-        rendered_message = self._prepare_assistant_message_for_render(message)
-        prompt_delta, completed_delta = self._render_assistant_turn_deltas(rendered_message, tools=tools)
-        assistant_header = "]~b]ai\n"
-        think_open = "<think>\n"
-        if prompt_delta != assistant_header + think_open or not completed_delta.startswith(assistant_header):
-            raise ValueError("Continuous Token MiniMax assistant scaffold does not match the supported protocol")
-
-        completed_body = completed_delta[len(assistant_header) :]
-        if completed_body.startswith(think_open):
-            assistant_text = completed_body[len(think_open) :]
-        else:
-            # The full-history template omits an empty reasoning block, but a
-            # rollout prompt has already emitted ``<think>\n``. The model must
-            # therefore generate the closing side before its visible answer.
-            assistant_text = "</think>\n\n" + completed_body
-        assistant_token_ids = normalize_token_ids(self.tokenizer.encode(assistant_text, add_special_tokens=False))
-        if not assistant_token_ids:
-            raise ValueError("Continuous Token assistant encoding produced an empty token-id suffix")
-        return self._normalize_assistant_token_ids(assistant_token_ids, message)
-
-    def _prepare_assistant_message_for_render(self, message: dict[str, Any]) -> dict[str, Any]:
-        if isinstance(message.get("reasoning_content"), str):
-            return message
-        content = message.get("content")
-        content_is_text_blocks = isinstance(content, list) and all(
-            isinstance(block, dict) and block.get("type") == "text" for block in content
-        )
-        if content_is_text_blocks:
-            content_text = "".join(str(block.get("text", "")) for block in content)
-        elif isinstance(content, str):
-            content_text = content
-        else:
-            return message
-        if not content_text.startswith("<think>") or "</think>" not in content_text:
-            return message
-
-        reasoning_content, answer_content = content_text[len("<think>") :].split("</think>", 1)
-        rendered_message = dict(message)
-        rendered_message["reasoning_content"] = reasoning_content
-        rendered_message["content"] = (
-            [{"type": "text", "text": answer_content}] if content_is_text_blocks else answer_content
-        )
-        return rendered_message
-
-    def _assistant_terminator_ids(self, message: dict[str, Any]) -> set[int]:
-        del message
-        return {self._eos_id}
-
     def _merge_non_assistant_token_ids(
         self, runtime_token_ids: list[int], appended_token_ids: list[int]
     ) -> MergeResult:
@@ -973,45 +695,6 @@ class GLMContinuousTokenBuilder(ContinuousTokenBuilder):
         self._user_id = require_token_id(tokenizer, "<|user|>")
         self._ambiguous_boundary_ids = {self._observation_id, self._user_id}
 
-    def _prepare_assistant_message_for_render(self, message: dict[str, Any]) -> dict[str, Any]:
-        enable_thinking = self.chat_template_kwargs.get("enable_thinking")
-        if isinstance(message.get("reasoning_content"), str):
-            if enable_thinking is not False:
-                return message
-            rendered_message = dict(message)
-            rendered_message["reasoning_content"] = ""
-            return rendered_message
-
-        content = message.get("content")
-        content_is_text_blocks = isinstance(content, list) and all(
-            isinstance(block, dict) and block.get("type") == "text" for block in content
-        )
-        if content_is_text_blocks:
-            content_text = "".join(str(block.get("text", "")) for block in content)
-        elif isinstance(content, str):
-            content_text = content
-        else:
-            if enable_thinking is not False:
-                return message
-            rendered_message = dict(message)
-            rendered_message["reasoning_content"] = ""
-            return rendered_message
-        if not content_text.startswith("<think>") or "</think>" not in content_text:
-            if enable_thinking is not False:
-                return message
-            rendered_message = dict(message)
-            rendered_message["reasoning_content"] = ""
-            return rendered_message
-
-        reasoning_content, answer_content = content_text[len("<think>") :].split("</think>", 1)
-        rendered_message = dict(message)
-        rendered_message["reasoning_content"] = reasoning_content if enable_thinking is not False else ""
-        answer_content = answer_content.lstrip("\n")
-        rendered_message["content"] = (
-            [{"type": "text", "text": answer_content}] if content_is_text_blocks else answer_content
-        )
-        return rendered_message
-
     def _merge_non_assistant_token_ids(
         self, runtime_token_ids: list[int], appended_token_ids: list[int]
     ) -> MergeResult:
@@ -1027,53 +710,6 @@ class GLMContinuousTokenBuilder(ContinuousTokenBuilder):
             removed_prefix_token_count=removed_prefix_token_count,
         )
 
-    def tokenize_assistant_message(
-        self,
-        message: dict[str, Any],
-        *,
-        tools: list[dict[str, Any]] | None = None,
-        previous_messages: list[dict[str, Any]] | None = None,
-    ) -> list[int]:
-        """Handle GLM's empty-reasoning scaffold without rewriting the prompt."""
-        del previous_messages
-        if message.get("role") != "assistant":
-            raise ValueError(
-                f"Continuous Token assistant encoding requires role='assistant', got {message.get('role')!r}"
-            )
-        rendered_message = self._prepare_assistant_message_for_render(message)
-        prompt_delta, completed_delta = self._render_assistant_turn_deltas(rendered_message, tools=tools)
-        assistant_header = "<|assistant|>"
-        if not prompt_delta.startswith(assistant_header) or not completed_delta.startswith(assistant_header):
-            raise ValueError("Continuous Token GLM assistant scaffold does not match the supported protocol")
-
-        if completed_delta.startswith(prompt_delta):
-            assistant_text = completed_delta[len(prompt_delta) :]
-        elif prompt_delta == assistant_header + "<think>":
-            # With empty reasoning the history template emits only </think>, so
-            # the completed render is not prefixed by the generation prompt's
-            # opening <think>. That opening token is already in runtime tokens.
-            assistant_text = completed_delta[len(assistant_header) :]
-        else:
-            raise ValueError("Continuous Token GLM completed assistant turn does not extend its generation scaffold")
-
-        assistant_token_ids = normalize_token_ids(self.tokenizer.encode(assistant_text, add_special_tokens=False))
-        if not assistant_token_ids:
-            raise ValueError("Continuous Token assistant encoding produced an empty token-id suffix")
-        return self._normalize_assistant_token_ids(assistant_token_ids, message)
-
-    def _normalize_assistant_token_ids(
-        self,
-        assistant_token_ids: list[int],
-        message: dict[str, Any],
-    ) -> list[int]:
-        # GLM templates do not terminate assistant text with EOS. Tool-call
-        # generation instead stops on <|observation|>, which is also the opening
-        # boundary of the following tool response.
-        normalized_ids = list(assistant_token_ids)
-        if message.get("tool_calls") and (not normalized_ids or normalized_ids[-1] != self._observation_id):
-            normalized_ids.append(self._observation_id)
-        return normalized_ids
-
 
 class Gemma4ContinuousTokenBuilder(ContinuousTokenBuilder):
     """Gemma4 tool-response boundary handling."""
@@ -1081,84 +717,6 @@ class Gemma4ContinuousTokenBuilder(ContinuousTokenBuilder):
     def __init__(self, tokenizer: Any, **kwargs: Any):
         super().__init__(tokenizer, **kwargs)
         self._tool_response_id = require_token_id(tokenizer, "<|tool_response>")
-        self._turn_id = require_token_id(tokenizer, "<turn|>")
-        self._tool_call_id = require_token_id(tokenizer, "<tool_call|>")
-
-    def tokenize_assistant_message(
-        self,
-        message: dict[str, Any],
-        *,
-        tools: list[dict[str, Any]] | None = None,
-        previous_messages: list[dict[str, Any]] | None = None,
-    ) -> list[int]:
-        """Reconstruct a Gemma4 response behind its thinking-mode scaffold."""
-        del previous_messages
-        if message.get("role") != "assistant":
-            raise ValueError(
-                f"Continuous Token assistant encoding requires role='assistant', got {message.get('role')!r}"
-            )
-
-        rendered_message = dict(message)
-        reasoning = next(
-            (
-                value
-                for key in ("thinking", "reasoning_content", "reasoning")
-                if isinstance((value := message.get(key)), str)
-            ),
-            "",
-        )
-        content = _stringify_tool_content(message.get("content", ""))
-        thought_open = "<|channel>thought\n"
-        thought_close = "<channel|>"
-        if content.startswith(thought_open) and thought_close in content:
-            embedded_reasoning, content = content[len(thought_open) :].split(thought_close, 1)
-            if not reasoning:
-                reasoning = embedded_reasoning
-            rendered_message["content"] = content
-
-        prompt_delta, completed_delta = self._render_assistant_turn_deltas(rendered_message, tools=tools)
-        assistant_header = "<|turn>model\n"
-        if not prompt_delta.startswith(assistant_header) or not completed_delta.startswith(assistant_header):
-            raise ValueError("Continuous Token Gemma4 assistant scaffold does not match the supported protocol")
-        completed_body = completed_delta[len(assistant_header) :]
-        prompt_scaffold = prompt_delta[len(assistant_header) :]
-        empty_thought_scaffold = thought_open + thought_close
-
-        if prompt_scaffold == empty_thought_scaffold:
-            # Standard 26B/31B non-thinking prompts already contain the complete
-            # empty thought channel, so only the visible completion is generated.
-            assistant_text = completed_body
-        elif not prompt_scaffold:
-            # Thinking prompts and E4B prompts end immediately after the model
-            # header. E4B may already serialize reasoning for tool calls; other
-            # templates drop it from history and require reconstruction here.
-            if completed_body.startswith(thought_open):
-                assistant_text = completed_body
-            elif reasoning or self.chat_template_kwargs.get("enable_thinking", False):
-                assistant_text = thought_open + reasoning + thought_close + completed_body
-            else:
-                assistant_text = completed_body
-        else:
-            raise ValueError("Continuous Token Gemma4 generation prompt has an unsupported scaffold")
-
-        assistant_token_ids = normalize_token_ids(self.tokenizer.encode(assistant_text, add_special_tokens=False))
-        if not assistant_token_ids:
-            raise ValueError("Continuous Token assistant encoding produced an empty token-id suffix")
-        return self._normalize_assistant_token_ids(assistant_token_ids, message)
-
-    def _normalize_assistant_token_ids(
-        self,
-        assistant_token_ids: list[int],
-        message: dict[str, Any],
-    ) -> list[int]:
-        terminator_id = self._tool_call_id if message.get("tool_calls") else self._turn_id
-        for index, token_id in enumerate(assistant_token_ids):
-            if token_id == terminator_id:
-                return list(assistant_token_ids[: index + 1])
-        raise ValueError(
-            "Continuous Token Gemma4 assistant token-id suffix does not contain the expected terminator "
-            f"{terminator_id}; tail={assistant_token_ids[-16:]}"
-        )
 
     def _should_fuse_generation_prompt_with_last_group(self) -> bool:
         # Gemma4's generation scaffold depends on the preceding message type and
@@ -1427,116 +985,12 @@ class DeepSeekContinuousTokenBuilder(ContinuousTokenBuilder):
 
     def __init__(self, tokenizer: Any, **kwargs: Any):
         super().__init__(tokenizer, **kwargs)
-        if "enable_thinking" in self.chat_template_kwargs and "thinking" not in self.chat_template_kwargs:
-            self.chat_template_kwargs = {
-                **self.chat_template_kwargs,
-                "thinking": self.chat_template_kwargs["enable_thinking"],
-            }
         # EOS is the only token guaranteed across V2/V3/R1
         self._eos_id = require_token_id(tokenizer, self._EOS_TOKEN)
         # V3/R1-specific tokens — lookup but tolerate absence (V2-Lite has none)
         self._bos_id = self._optional_token_id(tokenizer, self._BOS_TOKEN)
         self._user_id = self._optional_token_id(tokenizer, self._USER_TOKEN)
         self._assistant_id = self._optional_token_id(tokenizer, self._ASSISTANT_TOKEN)
-
-    def tokenize_assistant_message(
-        self,
-        message: dict[str, Any],
-        *,
-        tools: list[dict[str, Any]] | None = None,
-        previous_messages: list[dict[str, Any]] | None = None,
-    ) -> list[int]:
-        """Handle both classic DeepSeek and V3.1 thinking scaffolds."""
-        if message.get("role") != "assistant":
-            raise ValueError(
-                f"Continuous Token assistant encoding requires role='assistant', got {message.get('role')!r}"
-            )
-
-        if previous_messages and previous_messages[-1].get("role") == "tool":
-            return self._tokenize_assistant_after_tool(message, tools=tools)
-
-        prompt_delta, completed_delta = self._render_assistant_turn_deltas(message, tools=tools)
-        if completed_delta.startswith(prompt_delta):
-            assistant_text = completed_delta[len(prompt_delta) :]
-        else:
-            think_open = "<think>"
-            think_close = "</think>"
-            completed_prefix = prompt_delta[: -len(think_open)] + think_close
-            if not prompt_delta.endswith(think_open) or not completed_delta.startswith(completed_prefix):
-                raise ValueError(
-                    "Continuous Token DeepSeek completed assistant turn does not extend a supported generation scaffold"
-                )
-
-            reasoning = next(
-                (value for key in ("reasoning_content", "reasoning") if isinstance((value := message.get(key)), str)),
-                "",
-            )
-            content = _stringify_tool_content(message.get("content", ""))
-            if not reasoning and content.startswith(think_open) and think_close in content:
-                reasoning = content[len(think_open) :].split(think_close, 1)[0]
-            # V3.1's completed render already contains the closing tag and its
-            # visible answer/tool call, but drops reasoning. The generation
-            # prompt contains the opening tag, so put only the missing reasoning
-            # in front of that completed suffix.
-            assistant_header = completed_prefix[: -len(think_close)]
-            assistant_text = reasoning + completed_delta[len(assistant_header) :]
-
-        assistant_token_ids = normalize_token_ids(self.tokenizer.encode(assistant_text, add_special_tokens=False))
-        if not assistant_token_ids:
-            raise ValueError("Continuous Token assistant encoding produced an empty token-id suffix")
-        return self._normalize_assistant_token_ids(assistant_token_ids, message)
-
-    def _tokenize_assistant_after_tool(
-        self,
-        message: dict[str, Any],
-        *,
-        tools: list[dict[str, Any]] | None,
-    ) -> list[int]:
-        """Encode the direct continuation used after a DeepSeek tool result."""
-        synthetic_tool_message = {
-            "role": "tool",
-            "content": "continuous token synthetic tool response",
-            "tool_call_id": "continuous_token_call_0",
-            "name": _DUMMY_TOOL_NAME,
-        }
-        synthetic_context = [
-            _SYNTHETIC_SYSTEM_MESSAGE,
-            _SYNTHETIC_USER_MESSAGE,
-            self._synthetic_assistant_for_tools([synthetic_tool_message]),
-            synthetic_tool_message,
-        ]
-        prompt_text = self._render_text(synthetic_context, add_generation_prompt=True, tools=tools)
-        completed_text = self._render_text(
-            [*synthetic_context, message],
-            add_generation_prompt=False,
-            tools=tools,
-        )
-        if not completed_text.startswith(prompt_text):
-            raise ValueError(
-                "Continuous Token DeepSeek post-tool assistant encoding requires the generation prompt "
-                "to prefix the completed turn"
-            )
-        assistant_text = completed_text[len(prompt_text) :]
-        assistant_token_ids = normalize_token_ids(self.tokenizer.encode(assistant_text, add_special_tokens=False))
-        if not assistant_token_ids:
-            raise ValueError("Continuous Token assistant encoding produced an empty token-id suffix")
-        return self._normalize_assistant_token_ids(assistant_token_ids, message)
-
-    def _assistant_terminator_ids(self, message: dict[str, Any]) -> set[int]:
-        del message
-        return {self._eos_id}
-
-    def _synthetic_assistant_for_tools(
-        self,
-        tool_messages: list[dict[str, Any]],
-    ) -> dict[str, Any]:
-        synthetic_assistant = super()._synthetic_assistant_for_tools(tool_messages)
-        # DeepSeek V2/V3 templates concatenate arguments directly and therefore
-        # require serialized JSON rather than the dict accepted by ChatML-style
-        # templates.
-        for tool_call in synthetic_assistant["tool_calls"]:
-            tool_call["function"]["arguments"] = "{}"
-        return synthetic_assistant
 
     @staticmethod
     def _optional_token_id(tokenizer: Any, token: str) -> int | None:
@@ -1707,27 +1161,6 @@ class VLContinuousTokenMixin:
             tools=tools,
         )
 
-    def _render_text(
-        self,
-        messages: list[dict[str, Any]],
-        *,
-        add_generation_prompt: bool,
-        tools: list[dict[str, Any]] | None = None,
-    ) -> str:
-        template_kwargs = dict(self.chat_template_kwargs)
-        if tools:
-            template_kwargs["tools"] = tools
-        rendered = apply_chat_template(
-            self.processor,
-            _copy_messages_for_template(messages),
-            tokenize=False,
-            add_generation_prompt=add_generation_prompt,
-            **template_kwargs,
-        )
-        if not isinstance(rendered, str):
-            raise TypeError(f"Expected processor chat template to render str, got {type(rendered).__name__}")
-        return rendered
-
     def build_initial_tokens(
         self,
         messages: list[dict[str, Any]],
@@ -1881,44 +1314,6 @@ class MiniMaxVLContinuousTokenBuilder(VLContinuousTokenMixin, MiniMaxText01Conti
             raise ValueError("Continuous Token requires messages before a generation prompt")
         return list(self._vl_scaffold_ids)
 
-    def tokenize_assistant_message(
-        self,
-        message: dict[str, Any],
-        *,
-        tools: list[dict[str, Any]] | None = None,
-        previous_messages: list[dict[str, Any]] | None = None,
-    ) -> list[int]:
-        """Extract the VL assistant body after its unconditional scaffold."""
-        del tools, previous_messages
-        if message.get("role") != "assistant":
-            raise ValueError(
-                f"Continuous Token assistant encoding requires role='assistant', got {message.get('role')!r}"
-            )
-        rendered_message = self._prepare_vl_assistant_message(message)
-        synthetic_prompt = [_SYNTHETIC_SYSTEM_MESSAGE, _SYNTHETIC_USER_MESSAGE]
-        context_ids = self._render_tokens(synthetic_prompt, add_generation_prompt=False, tools=None)
-        prompt_ids = self._render_tokens(synthetic_prompt, add_generation_prompt=True, tools=None)
-        completed_ids = self._render_tokens(
-            [*synthetic_prompt, rendered_message],
-            add_generation_prompt=False,
-            tools=None,
-        )
-        if prompt_ids[: len(context_ids)] != context_ids or completed_ids[: len(context_ids)] != context_ids:
-            raise ValueError("Continuous Token MiniMax-VL assistant renders do not preserve the fixed context")
-        if prompt_ids[len(context_ids) :] != self._vl_scaffold_ids:
-            raise ValueError("Continuous Token MiniMax-VL generation prompt has an unsupported scaffold")
-        completed_delta = completed_ids[len(context_ids) :]
-        if completed_delta[: len(self._vl_scaffold_ids)] != self._vl_scaffold_ids:
-            raise ValueError("Continuous Token MiniMax-VL completed turn does not start with its assistant scaffold")
-        assistant_token_ids = completed_delta[len(self._vl_scaffold_ids) :]
-        if not assistant_token_ids:
-            raise ValueError("Continuous Token assistant encoding produced an empty token-id suffix")
-        return self._normalize_assistant_token_ids(assistant_token_ids, message)
-
-    @staticmethod
-    def _prepare_vl_assistant_message(message: dict[str, Any]) -> dict[str, Any]:
-        return _prepare_minimax_legacy_assistant_message(message)
-
 
 class Gemma4VLContinuousTokenBuilder(VLContinuousTokenMixin, Gemma4ContinuousTokenBuilder):
     """Gemma4 (unified) vision-language: Gemma4 ``<|tool_response>`` boundary handling
@@ -1969,21 +1364,6 @@ class KimiVLContinuousTokenBuilder(VLContinuousTokenMixin, ContinuousTokenBuilde
             audios=audios,
         )
 
-    def tokenize_assistant_message(
-        self,
-        message: dict[str, Any],
-        *,
-        tools: list[dict[str, Any]] | None = None,
-        previous_messages: list[dict[str, Any]] | None = None,
-    ) -> list[int]:
-        self._reject_tools(tools)
-        self._reject_structured_messages([message])
-        return super().tokenize_assistant_message(
-            message,
-            tools=None,
-            previous_messages=previous_messages,
-        )
-
     def merge_non_assistant_tokens(
         self,
         previous_messages: list[dict[str, Any]],
@@ -2011,19 +1391,6 @@ class KimiVLContinuousTokenBuilder(VLContinuousTokenMixin, ContinuousTokenBuilde
     ) -> list[int]:
         del tool_messages, previous_messages, tools, add_generation_prompt
         raise ValueError("Kimi-VL Continuous Token does not support structured tool responses")
-
-    def _normalize_assistant_token_ids(
-        self,
-        assistant_token_ids: list[int],
-        message: dict[str, Any],
-    ) -> list[int]:
-        im_end_id = self.tokenizer.convert_tokens_to_ids("<|im_end|>")
-        unk_token_id = getattr(self.tokenizer, "unk_token_id", None)
-        if isinstance(im_end_id, int) and im_end_id >= 0 and im_end_id != unk_token_id:
-            for index, token_id in enumerate(assistant_token_ids):
-                if token_id == im_end_id:
-                    return list(assistant_token_ids[: index + 1])
-        return super()._normalize_assistant_token_ids(assistant_token_ids, message)
 
 
 class DeepSeekVL2ContinuousTokenBuilder(DeepSeekContinuousTokenBuilder):
@@ -2106,18 +1473,14 @@ class DeepSeekVL2ContinuousTokenBuilder(DeepSeekContinuousTokenBuilder):
         self,
         messages: list[dict[str, Any]],
         images: list[Any],
-        add_generation_prompt: bool = True,
     ) -> list[int]:
         """Render messages through DeepseekVLV2Processor."""
-        conv, all_images = self._to_vl2_conversation(messages, images, add_generation_prompt)
-        # The official processor removes its final EOS in inference mode. A
-        # generation prompt needs that behavior, while a completed gold
-        # assistant turn must retain EOS as a generated continuation token.
+        conv, all_images = self._to_vl2_conversation(messages, images, add_generation_prompt=True)
         out = self.processor.__call__(
             conversations=conv,
             images=all_images,
             force_batchify=True,
-            inference_mode=add_generation_prompt,
+            inference_mode=True,
         )
         return normalize_token_ids(out.input_ids[0].tolist())
 
@@ -2134,42 +1497,7 @@ class DeepSeekVL2ContinuousTokenBuilder(DeepSeekContinuousTokenBuilder):
             images = self._extract_images_from_messages(messages)
         if tools:
             raise ValueError("DeepSeek-VL2 Continuous Token does not support tool schemas")
-        return self._render_via_processor(messages, images, add_generation_prompt=True)
-
-    def tokenize_assistant_message(
-        self,
-        message: dict[str, Any],
-        *,
-        tools: list[dict[str, Any]] | None = None,
-        previous_messages: list[dict[str, Any]] | None = None,
-    ) -> list[int]:
-        """Extract one assistant continuation through DeepseekVLV2Processor."""
-        del previous_messages
-        if message.get("role") != "assistant":
-            raise ValueError(
-                f"Continuous Token assistant encoding requires role='assistant', got {message.get('role')!r}"
-            )
-        if tools:
-            raise ValueError("DeepSeek-VL2 Continuous Token does not support tool schemas")
-        if message.get("tool_calls"):
-            raise ValueError("DeepSeek-VL2 Continuous Token does not support structured assistant tool calls")
-
-        synthetic_prompt = [_SYNTHETIC_USER_MESSAGE]
-        prompt_token_ids = self._render_via_processor(synthetic_prompt, [], add_generation_prompt=True)
-        completed_token_ids = self._render_via_processor(
-            [*synthetic_prompt, message],
-            [],
-            add_generation_prompt=False,
-        )
-        if completed_token_ids[: len(prompt_token_ids)] != prompt_token_ids:
-            raise ValueError(
-                "Continuous Token assistant encoding requires the processor generation prompt to be a token-id "
-                "prefix of the completed assistant turn"
-            )
-        assistant_token_ids = completed_token_ids[len(prompt_token_ids) :]
-        if not assistant_token_ids:
-            raise ValueError("Continuous Token assistant encoding produced an empty token-id suffix")
-        return self._normalize_assistant_token_ids(assistant_token_ids, message)
+        return self._render_via_processor(messages, images)
 
     def merge_non_assistant_tokens(
         self,
@@ -2192,7 +1520,7 @@ class DeepSeekVL2ContinuousTokenBuilder(DeepSeekContinuousTokenBuilder):
 
         # Always use full render + prefix diff (VL2 has no apply_chat_template)
         all_images = self._extract_images_from_messages(updated_messages)
-        full_token_ids = self._render_via_processor(updated_messages, all_images, add_generation_prompt=True)
+        full_token_ids = self._render_via_processor(updated_messages, all_images)
 
         prefix_len = len(runtime_token_ids)
         if full_token_ids[:prefix_len] != list(runtime_token_ids):
